move preview component to preview directory
