+++ web-ui/src/components/analyser/analyser.tsx
@@ -16,7 +16,7 @@ import { Lecture } from '@/types/lecture';
 import { useEffect, useState } from 'react';
 import apiClient, { ServerErrorResponse, ServerResponse } from '@/http';
 import { history } from 'umi';
-import Preview from '@/components/lecture/preview';
+import Preview from '@/components/lecture/preview/preview';
 import LectureProgress from './lecture-progress';
 import CourseSelector from './course-selector';
 import {
