+++ web-ui/src/components/lecture/preview/preview.less
